requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt
spacy
nltk
tensorflow
boto3
Pillow
pytesseract
bert-serving-server
bert-serving-client
pytorch-pretrained-bert
pyneuroner[cpu]
fuzzyset
fuzzywuzzy
pandas
